backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.0
pydantic==1.10.13